+++ src/python/test/test_nextflow.py
@@ -310,6 +310,23 @@ class TestRunNextflowApplet(DXTestCaseBuildNextflowApps):
         # are not printed
         self.assertNotIn("Launching", watched_run_output)
 
+    # @unittest.skipUnless(testutil.TEST_RUN_JOBS,
+    #                      'skipping tests that would run jobs')
+    # def test_dx_run_nextflow_with_unsupported_runtime_opts(self):
+    #     pipeline_name = "hello"
+    #     applet_dir = self.write_nextflow_applet_directory(pipeline_name, existing_nf_file_path=self.base_nextflow_nf)
+    #     applet_id = json.loads(run("dx build --nextflow --json " + applet_dir))["id"]
+    #     applet = dxpy.DXApplet(applet_id)
+
+    #     job = applet.run({
+    #                      "nextflow_run_opts": "-w user_workdir",
+    #     })
+        
+    #     job.wait_on_done()
+    #     job_desc = dxpy.describe(job.get_id())
+    #     self.assertEqual(job_desc["failureReason"], "AppError")
+    #     self.assertIn("Please remove workDir specification", job_desc["failureMessage"])
+
 if __name__ == '__main__':
     if 'DXTEST_FULL' not in os.environ:
         sys.stderr.write(